+++ apx_sdk/CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed `parse_http_signature` to `parse_http_signature_cavage`.
 - Return decoded signature string from `parse_http_signature`.
 - Renamed `HttpSignatureData.message` field to `base`.
+- Add header name to `HttpSignatureVerificationError::HeaderError` messages.
 
 ### Fixed